Refactor creating task buttons
