+++ package.json
@@ -231,7 +231,6 @@
         "react-textarea-autosize": "^7.1.0-1",
         "recompose": "^0.30.0",
         "redux": "^4.0.1",
-        "redux-action-watch": "^1.0.6",
         "redux-immutable": "^4.0.0",
         "redux-react-hook": "^3.4.0",
         "redux-saga": "^1.0.0",
