+++ src/modules/dashboard/components/ColonyHome/ColonyHome.css.d.ts
@@ -1,7 +1,6 @@
 export const wideButton: string;
 export const main: string;
 export const colonyInfo: string;
-export const interactiveBar: string;
 export const metaContainer: string;
 export const content: string;
 export const noTasks: string;
