+++ src/modules/dashboard/components/ColonyHome/ColonyTasks/ColonyTasks.css.d.ts
@@ -1,5 +1,2 @@
-export const newTaskButtonContainer: string;
-export const newTaskSpinnerContainer: string;
-export const newTaskButton: string;
 export const noTaskAddition: string;
 export const taskList: string;
