+++ src/modules/dashboard/components/ColonyHome/ColonyTasks/ColonyTasks.tsx
@@ -1,18 +1,11 @@
-import React, { useCallback, useEffect, useState } from 'react';
-import { defineMessages, FormattedMessage } from 'react-intl';
-import { subscribeActions as subscribeToReduxActions } from 'redux-action-watch/lib/actionCreators';
-import { useDispatch } from 'redux-react-hook';
-import throttle from 'lodash/throttle';
+import React from 'react';
 
 import { Address } from '~types/index';
-import { ActionTypes } from '~redux/index';
-import { mergePayload } from '~utils/actions';
-import { ActionButton } from '~core/Button';
-import Icon from '~core/Icon';
-import { SpinnerLoader } from '~core/Preloaders';
 import { useColonyTasksQuery } from '~data/index';
 
 import TaskList from '../../TaskList';
+import NewTaskButton from './NewTaskButton';
+
 import styles from './ColonyTasks.css';
 
 interface Props {
@@ -23,117 +16,26 @@ interface Props {
   filteredDomainId: number;
   showEmptyState?: boolean;
   canMintTokens?: boolean;
+  onCreateTask: () => void;
+  isCreatingTask: boolean;
 }
-
-const MSG = defineMessages({
-  newTask: {
-    id: 'dashboard.ColonyTasks.newTask',
-    defaultMessage: 'Create a new task',
-  },
-  newTaskDescription: {
-    id: 'dashboard.ColonyTasks.newTaskDescription',
-    defaultMessage: 'Create a new task',
-  },
-  noTasks: {
-    id: 'dashboard.ColonyTasks.noTasks',
-    defaultMessage: `It looks like you don't have any tasks.
-      Visit your colonies to find a task to work on.`,
-  },
-  welcomeToColony: {
-    id: 'dashboard.ColonyTasks.welcomeToColony',
-    defaultMessage: `Welcome to {colonyNameExists, select,
-      true {{colonyName}}
-      other {the Colony}
-    }!`,
-  },
-  creatingTask: {
-    id: 'dashboard.ColonyTasks.creatingTask',
-    defaultMessage: 'Creating your task...',
-  },
-  noCurrentlyOpenTasks: {
-    id: 'dashboard.TaskList.noCurrentlyOpenTasks',
-    defaultMessage: 'It looks like there are no open tasks right now.',
-  },
-});
 
 const displayName = 'dashboard.ColonyTasks';
 
-const NewTaskButton = ({
-  onClick,
-  disabled,
-  loading,
-}: {
-  onClick: Function;
-  disabled: boolean;
-  loading: boolean;
-}) => {
-  const [iconName, setIconName] = useState('empty-task');
-  if (disabled) {
-    return null;
-  }
-  if (loading) {
-    return (
-      <div className={styles.newTaskSpinnerContainer}>
-        <SpinnerLoader
-          appearance={{ theme: 'primary', size: 'massive' }}
-          loadingText={MSG.creatingTask}
-        />
-      </div>
-    );
-  }
-  return (
-    /*
-     * Ordinarily this wouldn't be necessary, but we can't use <button>
-     * because of the style requirements.
-     */
-    <div className={styles.newTaskButtonContainer}>
-      <Icon
-        className={styles.newTaskButton}
-        name={iconName}
-        onMouseEnter={() => setIconName('active-task')}
-        onMouseLeave={() => setIconName('empty-task')}
-        onClick={throttle(onClick as any, 2000)}
-        title={MSG.newTask}
-        viewBox="0 0 132 132"
-      />
-      <FormattedMessage tagName="p" {...MSG.newTaskDescription} />
-    </div>
-  );
-};
-
 const ColonyTasks = ({
   canCreateTask,
+  canMintTokens,
   colonyAddress,
   filterOption,
   filteredDomainId,
   isInRecoveryMode,
+  isCreatingTask,
+  onCreateTask,
   showEmptyState = true,
-  canMintTokens,
 }: Props) => {
-  const [isTaskBeingCreated, setIsTaskBeingCreated] = useState(false);
-  const dispatch = useDispatch();
-
-  /*
-   * This needs to return the `subscribeToReduxActions` function, since that returns an
-   * unsubscriber, and that gets called when the component is unmounted
-   */
-  useEffect(
-    () =>
-      subscribeToReduxActions(dispatch)({
-        [ActionTypes.TASK_CREATE]: () => setIsTaskBeingCreated(true),
-        [ActionTypes.TASK_CREATE_ERROR]: () => setIsTaskBeingCreated(false),
-      }),
-    [dispatch],
-  );
-
   const { data } = useColonyTasksQuery({
     variables: { address: colonyAddress },
   });
-
-  const transform = useCallback(
-    mergePayload({ colonyAddress, domainId: filteredDomainId }),
-    [filteredDomainId, colonyAddress],
-  );
 
   if (!data) {
     return null;
@@ -151,14 +53,10 @@ const ColonyTasks = ({
     ((canCreateTask && showEmptyState) || (canCreateTask && !canMintTokens))
   ) {
     return (
-      <ActionButton
-        button={NewTaskButton}
+      <NewTaskButton
         disabled={isInRecoveryMode}
-        error={ActionTypes.TASK_CREATE_ERROR}
-        submit={ActionTypes.TASK_CREATE}
-        success={ActionTypes.TASK_CREATE_SUCCESS}
-        transform={transform}
-        loading={isTaskBeingCreated}
+        loading={isCreatingTask}
+        onClick={onCreateTask}
       />
     );
   }
