+++ src/modules/dashboard/components/ColonyHome/ColonyTasks/NewTaskButton.css.d.ts
@@ -1,0 +1,3 @@
+export const newTaskButtonContainer: string;
+export const newTaskSpinnerContainer: string;
+export const newTaskButton: string;
