+++ src/modules/dashboard/components/ColonyHome/TabContribute/TabContribute.css
@@ -1,0 +1,6 @@
+.interactiveBar {
+  display: flex;
+  justify-content: space-between;
+  margin-top: 12px;
+  margin-bottom: 10px;
+}
