+++ src/modules/dashboard/components/ColonyHome/TabContribute/TabContribute.css.d.ts
@@ -1,0 +1,1 @@
+export const interactiveBar: string;
